refactor: drop canvas compat shims, import @/lib/canvas directly
Remove the editorOps/elementConfigs re-export shims and inline the
@/lib/canvas imports in AttributeBadge and CharactersPicker.
