+++ app/components/canvas/elements/CharactersPicker.tsx
@@ -9,11 +9,11 @@ import {
 	DropdownMenuItem,
 	DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
+import { setNodeAttrs } from "@/lib/canvas/editorOps";
+import type { CanvasContentElement } from "@/lib/canvas/types";
 import { useConfig } from "@/lib/config/ConfigProvider";
 import { useProjectStore } from "@/lib/project/store";
-import type { CanvasContentElement } from "@/lib/canvas/types";
 import { getElementCharacterNames } from "../utils/characters";
-import { setNodeAttrs } from "../utils/editorOps";
 import { CharacterPill } from "./CharacterBadge";
 
 function useProjectCharacterNames(): string[] {
