--- app/components/canvas/utils/editorOps.ts
@@ -1,3 +1,0 @@
-// Compat re-export: keeps files touched by open PR #263 unchanged.
-// Fold into direct @/lib/canvas/editorOps imports once it lands.
-export { setNodeAttrs } from "@/lib/canvas/editorOps";
